requirements.txt
Flask==3.0.0
python-docx==1.1.0
beautifulsoup4==4.12.2
requests==2.31.0
google-generativeai==0.3.2
python-dotenv==1.0.0
Werkzeug==3.0.1
docx2pdf==0.1.8
